requirements.txt
blinker==1.6.2
certifi>=2024.2.0
selenium>=4.12.0
selenium-wire>=5.1.0
undetected-chromedriver>=3.5.0
gspread>=5.12.0
oauth2client>=4.1.3
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
requests>=2.31.0
colorama>=0.4.6
tenacity>=8.2.0
tqdm>=4.66.0
setuptools>=65.0.0
urllib3>=2.0.0
